Default values for ppgmmga package

Set or retrieve default values to be used by the ppgmmga package.

ppgmmga.options(…)

Arguments

…
A single character vector, or a named list with components. In the one argument case, the form name = value can be used to change a single option. In the multiple arguments case, the form list(name1 = value1, name2 = value2) can be used to change several arguments. If no arguments are provided, then the function returns all the current options. For the available options see the Details section below.

Details

This function can be used to set or retrieve the values to be used by the ppgmmga package.

The function globally sets the arguments for the current session of R. The default options are restored with a new R session. To temporarily change the options for a single call to ppgmmga function, look at options argument in ppgmmga.

Available options are:

modelNames

A string specifying the GMM to fit. See mclustModelNames for the available models.

G

An integer value or a vector of integer values specifying the number of mixture components. If more than a single value is provided, the best model is selected using the BIC criterion. By default G = 1:9.

initMclust

A string specifying the type of initialisation to be used for the EM algorithm. See mclust.options for more details.

popSize

The GA population size. By default popSize = 100.

pcrossover

The probability of crossover. By default pcrossover = 0.8.

pmutation

The probability of mutation. By default pmutation = 0.1.

maxiter

An integer value specifying the maximum number of iterations before stopping the GA. By default maxiter = 1000.

run

An integer value indicating the number of generations without improvment in the best value of fitness fuction. run = 100.

selection

An R function performing the selection genetic operator. See ga_Selection for details. By default selection = gareal_lsSelection.

crossover

An R function performing the crossover genetic operator. See ga_Crossover for details. By default crossover = gareal_laCrossover.

mutation

An R function performing the mutation genetic operator. See ga_Mutation for details. By default mutation = gareal_raMutation.

parallel

A logical value specifying whether or not GA should be run in parallel. By default parallel = FALSE.

numIslands

An integer value specifying the number of islands to be used in the Island Genetic Algorithm. By default numIslands = 4.

migrationRate

A value specifying the fraction of migration between islands. By default migrationRate = 0.1.

migrationInterval

An integer values specifying the number of generations to run before each migration. By default migrationInterval = 10.

optim

A logical value specifying whether or not a local search should be performed. By default optim = TRUE.

optimPoptim

A value specifying the probability a local search is performed at each GA generation. By default optimPoptim = 0.05.

optimPressel

A value in \([0,1]\) specifying the pressure selection. Values close to 1 tend to assign higher selection probabilities to solutions with higher fitness, whereas values close to 0 tend to assign equal selection probability to any solution. By default optimPressel = 0.5.

optimMethod

A string specifying the general-purpose optimisation method to be used for local search. See optim for the available algorithms. By default optimMethod = "L-BFGS-B".

optimMaxit

An integer value specifying the number of iterations for the local search algorithm. By default optimMaxit = 100.

orth

A string specifying the method employed to orthogonalise the matrix basis. Available methods are the QR decomposition "QR", and the Singular Value Decomposition "SVD". By default orth = "QR".

classPlotSymbols

A vector whose entries are either integers corresponding to graphics symbols or single characters for indicating classifications when plotting data. Classes are assigned symbols in the given order.

classPlotColors

A vector whose entries correspond to colors for indicating classifications when plotting data. Classes are assigned colors in the given order.

For more details about options related to Gaussian mixture modelling see densityMclust, and for those related to genetic algorithms see ga and gaisl.

References

Scrucca, L., Fop, M., Murphy, T. B., & Raftery, A. E. (2016) mclust 5: Clustering, classification and density estimation using gaussian finite mixture models. The R journal, 8(1), 205-233. https://journal.r-project.org/archive/2016-1/scrucca-fop-murphy-etal.pdf

Scrucca, L. (2013) GA: A Package for Genetic Algorithms in R. Journal of Statistical Software, 53(4), 1-37. http://www.jstatsoft.org/v53/i04/

Scrucca, L. (2017) On some extensions to GA package: hybrid optimisation, parallelisation and islands evolution. The R Journal, 9/1, 187-206. https://journal.r-project.org/archive/2017/RJ-2017-008

Scrucca, L. and Serafini, A. (2019) Projection pursuit based on Gaussian mixtures and evolutionary algorithms. *Journal of Computational and Graphical Statistics*, DOI: 10.1080/10618600.2019.1598871

See also

Examples

# NOT RUN {
ppgmmga.options()

# Print a single option
ppgmmga.options("popSize")

# Change (globally) an option
ppgmmga.options("popSize" = 10)
ppgmmga.options("popSize")
# }